tidy up and improve comments
had a little detour into preparing lists of snapshots to create and
passing it as a list to a single zfs invocation but that complicated the
script. while it might be more performant and would create atmoic
snapshots, I opted for the KISS principle .. keep it simple.

think about proper atomic snapshots again when zfs 8.0 and channel
programs are available.